refactor: Rename base controller property in Controller class
The property `$baseController` within the Controller class is renamed to **`$baseControllerName`**.

This aligns the property name with the new naming convention used in the bootstrapping process, improving clarity by explicitly indicating that the variable holds the string name of the base controller. The property is also updated in the constructor of helper objects.

Signed-off-by: chris <chris@ckvsoft.at>