cleanup(js): Remove commented-out, unused code block from menuscript.js
This commit cleans up the `menuscript.js` file by removing a large block of commented-out JavaScript code that was previously used for setting the margin of the main container based on the header height and for displaying and fading out a status message. This code is no longer necessary and was cluttering the file.

Signed-off-by: chris <chris@ckvsoft.at>